feat(test): add postgres and rabbitmq services to ci
Adds postgres and rabbitmq as services to the test workflow. This ensures that the tests run in an environment that is as close as possible to production.

This commit also introduces django-smart-env to manage the demo application configuration, making it more flexible and removing hardcoded values.

The tox and mypy configurations have been updated to support these changes.