feat: Homoglyph support for text search
This commit introduces homoglyph support for text search, which improves the quality of the search by considering characters that look similar but are different. This is achieved by using the rapidfuzz library, which provides a set of tools for fuzzy string matching.

The find_similar function has been refactored to support two different implementations. The second one, which is the one enabled by default, is the one that supports homoglyphs.

The ScanInfo and SearchInfo classes have been refactored to be more clear and to better represent the data.

The tests have been updated to reflect the changes in the find_similar function and to better test the new functionality.

The rapidfuzz library has been added to the dependencies.

The tests now run in parallel, which should speed up the test suite.

The pytest_configure function in tests/conftest.py has been updated to remove the DJANGO_SETTINGS_MODULE environment variable, which is no longer needed.

A test image has been removed.